--- a/include/AdePT/kernels/gammas.cuh
+++ b/include/AdePT/kernels/gammas.cuh
@@ -208,8 +208,7 @@ __global__ void TransportGammas(adept::TrackManager<Track> *gammas, Secondaries 
       // Check the cuts and deposit energy in this volume if needed
       double edep=0;
 
-      if (false) {
-      // if (APPLY_CUTS && elKinEnergy < theElCut) {
+      if (APPLY_CUTS && elKinEnergy < theElCut) {
         // Deposit the energy here and kill the secondary
         edep += elKinEnergy;
       }
@@ -223,9 +222,8 @@ __global__ void TransportGammas(adept::TrackManager<Track> *gammas, Secondaries 
         electron.dir.Set(dirSecondaryEl[0], dirSecondaryEl[1], dirSecondaryEl[2]);
       }
 
-      if (false) {
-      // if (APPLY_CUTS && 
-      //       (copcore::units::kElectronMassC2 < theGammaCut && posKinEnergy < theElCut)) {
+      if (APPLY_CUTS && 
+            (copcore::units::kElectronMassC2 < theGammaCut && posKinEnergy < theElCut)) {
         // Deposit: posKinEnergy + 2 * copcore::units::kElectronMassC2 and kill the secondary
         edep += posKinEnergy + 2 * copcore::units::kElectronMassC2;
       }
@@ -287,9 +285,8 @@ __global__ void TransportGammas(adept::TrackManager<Track> *gammas, Secondaries 
       // Check the cuts and deposit energy in this volume if needed
       double edep=0;
 
-      if (energyEl > LowEnergyThreshold) {
-      // if (energyEl > LowEnergyThreshold && 
-      //     (APPLY_CUTS && (energyEl > theElCut))) {
+      if (energyEl > LowEnergyThreshold && 
+          (APPLY_CUTS && (energyEl > theElCut))) {
         // Create a secondary electron and sample/compute directions.
         Track &electron = secondaries.electrons->NextTrack();
 
@@ -348,9 +345,8 @@ __global__ void TransportGammas(adept::TrackManager<Track> *gammas, Secondaries 
 
       double edep             = bindingEnergy;
       const double photoElecE = eKin - edep;
-      if (photoElecE > theLowEnergyThreshold) {
-      // if (photoElecE > theLowEnergyThreshold && 
-      //     (APPLY_CUTS && (photoElecE > theElCut))) {
+      if (photoElecE > theLowEnergyThreshold && 
+          (APPLY_CUTS && (photoElecE > theElCut))) {
         // Create a secondary electron and sample directions.
         Track &electron = secondaries.electrons->NextTrack();
         adept_scoring::AccountProduced(userScoring, /*numElectrons*/ 1, /*numPositrons*/ 0, /*numGammas*/ 0);