--- a/include/AdePT/kernels/gammas.cuh
+++ b/include/AdePT/kernels/gammas.cuh
@@ -258,7 +258,8 @@ __global__ void TransportGammas(adept::TrackManager<Track> *gammas, Secondaries 
                                     &dir,                  // Post-step point momentum direction
                                     nullptr,               // Post-step point polarization
                                     0,                     // Post-step point kinetic energy
-                                    0);                    // Post-step point charge
+                                    0,                     // Post-step point charge
+                                    0, -1);                // event and thread ID
       }
       
       // The current track is killed by not enqueuing into the next activeQueue.